fix: move typscript to deps. Move types and Commitizen to dev deps
This change moves typescript to deps to to fix a bug where typescript could not be resolved when
graphql-usage was installed globally. This change also cleans up dependencies for types and
Commitizen. --ignore-engines has also been added to the install step on CI to prevent Commitzen to
cause yarn to throw errors when using Node 8.